third_party/WebKit/LayoutTests/http/tests/feature-policy-experimental-features/vibrate-allowed-by-container-policy-relocate-and-no-reload-expected.txt

CONSOLE ERROR: line 4: Error while parsing the 'allow' attribute: 'vibrate' is an invalid feature name.
This is a testharness.js-based test.
FAIL Iframe navigated itself to a new origin, vibrate is disabled by container policy.assert_false: navigator.vibrate(): expected false got true
Harness: the test ran to completion.
